simplified and consolidated direct collocation constraint replaced ExtractGradientMatrix with math::autodiff_gradient version
update cmakelists
