+++ drake/systems/trajectory_optimization/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_library_with_exports(LIB_NAME drakeTrajectoryOptimization SOURCE_FILES
   direct_collocation.cc
-  direct_collocation_constraint.cc
     multiple_shooting.cc)
 target_link_libraries(drakeTrajectoryOptimization
   drakeOptimization
@@ -10,7 +9,6 @@ target_link_libraries(drakeTrajectoryOptimization
 drake_install_libraries(drakeTrajectoryOptimization)
 drake_install_headers(
   direct_collocation.h
-  direct_collocation_constraint.h
     multiple_shooting.h)
 drake_install_pkg_config_file(drake-trajectory-optimization
   TARGET drakeTrajectoryOptimization
@@ -21,6 +19,3 @@ drake_install_pkg_config_file(drake-trajectory-optimization
     drake-trajectories
     eigen3)
 
-if(BUILD_TESTING)
-  add_subdirectory(test)
-endif()
