--- drake/systems/trajectory_optimization/test/CMakeLists.txt
@@ -1,2 +1,0 @@
-drake_add_cc_test(direct_collocation_constraint_test)
-target_link_libraries(direct_collocation_constraint_test drakeTrajectoryOptimization)
